Rename `models.transformer_model` to `models.architectures`
