+++ curated_transformers/models/__init__.py
@@ -1,6 +1,4 @@
-from .hf_loader import build_hf_encoder_loader_v1
-from .scalar_weight import build_scalar_weight_v1
-from .transformer_model import (
+from .architectures import (
     build_albert_transformer_model_v1,
     build_bert_transformer_model_v1,
     build_camembert_transformer_model_v1,
@@ -8,4 +6,6 @@ from .transformer_model import (
     build_xlmr_transformer_model_v1,
     build_pytorch_checkpoint_loader_v1,
 )
+from .hf_loader import build_hf_encoder_loader_v1
+from .scalar_weight import build_scalar_weight_v1
 from .with_strided_spans import build_with_strided_spans_v1
