+++ curated_transformers/models/architectures.py
