+++ curated_transformers/tests/models/test_hf_model.py
@@ -5,6 +5,7 @@ import torch
 from torch.nn import Module
 
 from curated_transformers._compat import has_hf_transformers, transformers
+from curated_transformers.models.architectures import _pytorch_encoder
 from curated_transformers.models.hf_loader import build_hf_encoder_loader_v1
 from curated_transformers.models.pytorch.albert import AlbertEncoder
 from curated_transformers.models.pytorch.albert.config import AlbertConfig
@@ -12,7 +13,6 @@ from curated_transformers.models.pytorch.attention import AttentionMask
 from curated_transformers.models.pytorch.bert import BertConfig, BertEncoder
 from curated_transformers.models.pytorch.roberta.config import RobertaConfig
 from curated_transformers.models.pytorch.roberta.encoder import RobertaEncoder
-from curated_transformers.models.transformer_model import _pytorch_encoder
 
 
 @dataclass
